refactor(memory-wiki): own the shared import result type
Move the private result definition to the shared operation and migrate
plugin-local readers, removing the type-only back-edge to its caller.
No public barrel or executable import flow changes.

Proof: pnpm check:architecture passed (runtime and Madge cycles zero);
67 focused tests and extension production/test types passed.
