+++ extensions/memory-wiki/src/source-sync.ts
@@ -1,6 +1,6 @@
 // Memory Wiki plugin module implements source sync behavior.
 import type { OpenClawConfig } from "../api.js";
-import { syncMemoryWikiBridgeSources, type BridgeMemoryWikiResult } from "./bridge.js";
+import { syncMemoryWikiBridgeSources } from "./bridge.js";
 import {
   refreshMemoryWikiIndexesAfterImport,
   type RefreshMemoryWikiIndexesResult,
@@ -10,6 +10,7 @@ import {
   resolveMemoryWikiVaultMutationKey,
   withMemoryWikiVaultMutation,
 } from "./mutation-coordinator.js";
+import type { BridgeMemoryWikiResult } from "./source-import.js";
 import { syncMemoryWikiUnsafeLocalSources } from "./unsafe-local.js";
 import { initializeMemoryWikiVault } from "./vault.js";
